Applied coordinate picker popup for create and edit venue

+++ src/main/java/com/panchoriz/myapp/model/VenueDTO.java
@@ -19,8 +19,8 @@ public class VenueDTO implements BaseDTO {
 	private int courtType;
 	private int flooringType;
 	
-	private long longitude;
-	private long latitude;
+	private double longitude;
+	private double latitude;
 	
 	public String getVenueName() {
 		return venueName;
@@ -70,16 +70,16 @@ public class VenueDTO implements BaseDTO {
 	public void setFlooringType(int flooringType) {
 		this.flooringType = flooringType;
 	}
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	public String getVenueId() {
