+++ src/main/java/com/panchoriz/myapp/repositories/documents/VenueDocument.java
@@ -25,8 +25,8 @@ public class VenueDocument extends AbstractDocument<String> {
 	private int courtType;
 	private int flooringType;
 	
-	private long longitude;
-	private long latitude;
+	private double longitude;
+	private double latitude;
 	
 	public String getVenueName() {
 		return venueName;
@@ -76,16 +76,16 @@ public class VenueDocument extends AbstractDocument<String> {
 	public void setFlooringType(int flooringType) {
 		this.flooringType = flooringType;
 	}
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 }
